# openapi.json
{
    "openapi": "3.1.0",
    "info": {
      "title": "GPT Stripe Payment API",
      "version": "1.0.0",
      "description": "API for managing users, products, orders, and payments via Stripe."
    },
    "servers": [
      {
        "url": "https://gpt-stripe-store-main-for-nexus1e.vercel.app/",
        "description": "Production environment"
      },
      {
        "url": "https://gpt-stripe-store-dev.for-nexus1e.vercel.app/",
        "description": "Development environment"
      }
    ],
    "paths": {
...
      "/orders": {
        "post": {
          "summary": "Create a New Order",
          "operationId": "createOrder",
          "requestBody": {
            "description": "Order details",
            "required": true,
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/OrderCreate"
                }
              }
            }
          },
          "responses": {
            "201": {
              "description": "Order created successfully.",
              "content": {
                "application/json": {
                  "schema": {
                    "$ref": "#/components/schemas/Order"
                  }
                }
              }
            },
            "400": {
              "description": "Invalid order data.",
              "content": {
                "application/json": {
                  "schema": {
                    "$ref": "#/components/schemas/ErrorResponse"
                  }
                }
              }
            },
            "401": {
              "description": "Unauthorized.",
              "content": {
                "application/json": {
                  "schema": {
                    "$ref": "#/components/schemas/ErrorResponse"
                  }
                }
              }
            }
          },
          "security": [
            {
              "BearerAuth": []
            }
          ]
        },
        "get": {
          "summary": "List User's Orders",
          "operationId": "listUserOrders",
          "parameters": [
            {
              "name": "status",
              "in": "query",
              "required": false,
              "schema": {
                "type": "string",
                "enum": ["pending", "paid", "failed"]
              },
              "description": "Filter orders by status"
            },
            {
              "name": "page",
              "in": "query",
              "required": false,
              "schema": {
                "type": "integer",
                "default": 1
              },
              "description": "Page number for pagination"
            },
            {
              "name": "limit",
              "in": "query",
              "required": false,
              "schema": {
                "type": "integer",
                "default": 10
              },
              "description": "Number of orders per page"
            }
          ],
          "responses": {
            "200": {
              "description": "List of user orders.",
              "content": {
                "application/json": {
                  "schema": {
                    "type": "array",
                    "items": {
                      "$ref": "#/components/schemas/Order"
                    }
                  }
                }
              }
            },
            "401": {
              "description": "Unauthorized.",
              "content": {
                "application/json": {
                  "schema": {
                    "$ref": "#/components/schemas/ErrorResponse"
                  }
                }
              }
            }
          },
          "security": [
            {
              "BearerAuth": []
            }
          ]
        }
      },
...
      "/webhook/stripe": {
        "post": {
          "summary": "Handle Stripe Webhook Events",
          "operationId": "handleStripeWebhook",
          "requestBody": {
            "description": "Stripe webhook event payload",
            "required": true,
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "id": {
                      "type": "string",
                      "description": "Unique identifier for the webhook event"
                    },
                    "object": {
                      "type": "string",
                      "description": "Type of the webhook object"
                    },
                    "type": {
                      "type": "string",
                      "description": "Event type, e.g., 'payment_intent.succeeded'"
                    },
                    "data": {
                      "type": "object",
                      "description": "Event data payload"
                    },
                    "created": {
                      "type": "integer",
                      "description": "Timestamp when the event was created"
                    },
                    "livemode": {
                      "type": "boolean",
                      "description": "Indicates if the event was sent from live mode"
                    }
                  },
                  "required": ["id", "object", "type", "data", "created", "livemode"]
                }
              }
            }
          },
          "responses": {
            "200": {
              "description": "Webhook event processed successfully.",
              "content": {
                "application/json": {
                  "schema": {
                    "$ref": "#/components/schemas/WebhookResponse"
                  }
                }
              }
            },
            "400": {
              "description": "Invalid webhook payload.",
              "content": {
                "application/json": {
                  "schema": {
                    "$ref": "#/components/schemas/ErrorResponse"
                  }
                }
              }
            }
          }
        }
      },
      "/hasUserPaid/{orderId}": {
        "get": {
          "summary": "Check If User Has Paid for an Order",
          "operationId": "hasUserPaid",
          "parameters": [
            {
              "name": "orderId",
              "in": "path",
              "required": true,
              "schema": {
                "type": "string",
                "format": "uuid"
              },
              "description": "Unique ID of the order"
            }
          ],
          "responses": {
            "200": {
              "description": "User payment status retrieved.",
              "content": {
                "application/json": {
                  "schema": {
                    "$ref": "#/components/schemas/PaymentStatusResponse"
                  }
                }
              }
            },
            "404": {
              "description": "Order not found.",
              "content": {
                "application/json": {
                  "schema": {
                    "$ref": "#/components/schemas/ErrorResponse"
                  }
                }
              }
            },
            "401": {
              "description": "Unauthorized.",
              "content": {
                "application/json": {
                  "schema": {
                    "$ref": "#/components/schemas/ErrorResponse"
                  }
                }
              }
            }
          },
          "security": [
            {
              "BearerAuth": []
            }
          ]
        }
      },
...
    },
    "components": {
      "schemas": {
...
        "OrderCreate": {
          "type": "object",
          "properties": {
            "products": {
              "type": "array",
              "items": {
                "$ref": "#/components/schemas/OrderItemCreate"
              },
              "description": "List of products to include in the order."
            }
          },
          "required": ["products"]
        },
        "OrderItemCreate": {
          "type": "object",
          "properties": {
            "product_id": {
              "type": "string",
              "format": "uuid",
              "example": "prod_ABC123XYZ"
            },
            "quantity": {
              "type": "integer",
              "example": 2
            }
          },
          "required": ["product_id", "quantity"]
        },
        "Order": {
          "type": "object",
          "properties": {
            "order_id": {
              "type": "string",
              "format": "uuid",
              "example": "order_ABC123XYZ"
            },
            "user_id": {
              "type": "string",
              "format": "uuid",
              "example": "123e4567-e89b-12d3-a456-426614174000"
            },
            "total_amount": {
              "type": "number",
              "format": "float",
              "example": 99.98
            },
            "currency": {
              "type": "string",
              "example": "usd"
            },
            "status": {
              "type": "string",
              "enum": ["pending", "paid", "failed"],
              "example": "pending"
            },
            "stripe_payment_intent_id": {
              "type": "string",
              "example": "pi_ABC123XYZ"
            },
            "created_at": {
              "type": "string",
              "format": "date-time",
              "example": "2025-01-10T12:34:56Z"
            },
            "updated_at": {
              "type": "string",
              "format": "date-time",
              "example": "2025-01-10T12:34:56Z"
            },
            "order_items": {
              "type": "array",
              "items": {
                "$ref": "#/components/schemas/OrderItem"
              }
            }
          }
        },
        "OrderItem": {
          "type": "object",
          "properties": {
            "order_item_id": {
              "type": "string",
              "format": "uuid",
              "example": "order_item_ABC123XYZ"
            },
            "order_id": {
              "type": "string",
              "format": "uuid",
              "example": "order_ABC123XYZ"
            },
            "product_id": {
              "type": "string",
              "format": "uuid",
              "example": "prod_ABC123XYZ"
            },
            "quantity": {
              "type": "integer",
              "example": 2
            },
            "unit_price": {
              "type": "number",
              "format": "float",
              "example": 49.99
            },
            "total_price": {
              "type": "number",
              "format": "float",
              "example": 99.98
            },
            "created_at": {
              "type": "string",
              "format": "date-time",
              "example": "2025-01-10T12:34:56Z"
            },
            "updated_at": {
              "type": "string",
              "format": "date-time",
              "example": "2025-01-10T12:34:56Z"
            }
          }
        },
...
        "WebhookResponse": {
          "type": "object",
          "properties": {
            "status": {
              "type": "string",
              "example": "success"
            },
            "message": {
              "type": "string",
              "example": "Webhook event processed successfully."
            }
          }
        },
        "PaymentStatusResponse": {
          "type": "object",
          "properties": {
            "hasPaid": {
              "type": "boolean",
              "example": true
            }
          }
        },
...
        "ErrorResponse": {
          "type": "object",
          "properties": {
            "status": {
              "type": "string",
              "example": "error"
            },
            "message": {
              "type": "string",
              "example": "Detailed error message."
            }
          }
        }
      },
      "securitySchemes": {
        "BearerAuth": {
          "type": "http",
          "scheme": "bearer",
          "bearerFormat": "JWT",
          "description": "JWT Authorization header using the Bearer scheme. Example: \"Authorization: Bearer {token}\""
        }
      }
    },
...
  }